third_party/blink/web_tests/crypto/subtle/modify-importKey-data-during-normalization-expected.txt

Tests crypto.subtle.importKey() using a BufferSource that is modified during algorithm normalization

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing key...
Accessed name property
Corrupting keyData...
Exporting key...
PASS: Exported data should be [30112233445566778899aabbccddeeff] and was
Importing key (again)...
Importing second key...
PASS: Exported data should be [00002233445566778899aabbccddeeff] and was
PASS successfullyParsed is true

TEST COMPLETE
